Changed BiDiIFDSSOlver to support IDE, new version of BiDiIFDSSolver based on BiDiIDESolver
